% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhf_theme.R
\name{bhf_bar_plot}
\alias{bhf_bar_plot}
\title{bhf_bar_plot - wrapper function for bhf_theme to support legacy bhf_bar_plot function.
Equivalent to bhf_theme(line=TRUE,grid=FALSE)}
\usage{
bhf_bar_plot(textsize = 12)
}
\arguments{
\item{textsize}{Size of text (defaults to 12)}

\item{line}{Boolean for whether line elements should be drawn}

\item{grid}{Boolean for whether grid lines should be drawn}

\item{map}{Boolean for a map object}
}
\value{
ggplot2 style object
}
\description{
bhf_bar_plot - wrapper function for bhf_theme to support legacy bhf_bar_plot function.
Equivalent to bhf_theme(line=TRUE,grid=FALSE)
}
\examples{
theme <- bhf_bar_plot(line=TRUE, grid=FALSE)
}
