% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhf_theme.R
\name{bhf_colourcode}
\alias{bhf_colourcode}
\title{BHF Colours}
\usage{
bhf_colourcode(colour)
}
\arguments{
\item{colour}{list of names of BHF colours to return. Case insensitive}
}
\value{
BHF colour hex code
}
\description{
Returns the colour
}
\examples{
colour <- bhf_colours("dark grey")
colours <- bhf_colours(c("Bright Red", "Dark Red"))
}
