% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhf_theme.R
\name{bhf_cols}
\alias{bhf_cols}
\title{BHF Column Conversion}
\usage{
bhf_cols(...)
}
\arguments{
\item{column_names}{Input columns}
}
\value{
BHF colour columns or those columns found in input
}
\description{
Convert degrees Celsius temperatures to degrees Fahrenheit
}
\examples{
temp1 <- bhf_cols(c("Bright Red","Dark Red"));
}
