% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhf_theme.R
\name{bhf_line_plot}
\alias{bhf_line_plot}
\title{bhf_line_plot - wrapper function for bhf_theme to support legacy bhf_line_plot function
Equivalent to bhf_theme(line=TRUE)}
\usage{
bhf_line_plot(textsize = 12)
}
\arguments{
\item{textsize}{Size of text (defaults to 12)}

\item{line}{Boolean for whether line elements should be drawn}

\item{grid}{Boolean for whether grid lines should be drawn}

\item{map}{Boolean for a map object}
}
\value{
ggplot2 style object
}
\description{
bhf_line_plot - wrapper function for bhf_theme to support legacy bhf_line_plot function
Equivalent to bhf_theme(line=TRUE)
}
\examples{
theme <- bhf_line_plot(grid=TRUE)
}
