% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhf_theme.R
\name{bhf_pal}
\alias{bhf_pal}
\title{BHF Palettes}
\usage{
bhf_pal(palette = "reds", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Type of palette (defaults to reds)}

\item{reverse}{Reverses the palette order}
}
\value{
Palette object
}
\description{
Get a BHF palette
}
\examples{
temp1 <- bhf_pal("reds",reverse=TRUE)
}
