% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhf_theme.R
\name{bhf_palette}
\alias{bhf_palette}
\title{BHF Palettes
Returns a BHF palette from the standard roster}
\usage{
bhf_palette(palette = "bhf colours", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Type of palette - any of "bhf colours","red to yellow", "expanded colours", "imd decile", "map colours", "imd quintiles". Defaults to standard BHF colours}

\item{reverse}{Reverses the palette order}
}
\value{
Palette object
}
\description{
BHF Palettes
Returns a BHF palette from the standard roster
}
\examples{
temp1 <- bhf_pal("reds", reverse = TRUE)
}
