% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhf_theme.R
\docType{data}
\name{bhf_palettes}
\alias{bhf_palettes}
\title{BHF Column Conversion}
\format{
An object of class \code{list} of length 6.
}
\usage{
bhf_palettes
}
\arguments{
\item{column_names}{Input columns}
}
\value{
BHF colour columns or those columns found in input
}
\description{
BHF Column Conversion
}
\examples{
temp1 <- bhf_cols(c("Bright Red", "Dark Red"))
}
\keyword{datasets}
