% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhf_theme.R
\name{bhf_style}
\alias{bhf_style}
\title{BHF Style}
\usage{
bhf_style(bhf_brand, textsize = 25)
}
\arguments{
\item{bhf_brand}{Branding type}

\item{textsize}{Size of text (defaults to 25)}
}
\value{
ggplot2 style object
}
\description{
Defines the overall BHF plotting theme
}
\examples{
theme <- bhf_style("reds",reverse=TRUE)
}
