% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhf_theme.R
\name{bhf_theme}
\alias{bhf_theme}
\title{bhf_theme - The main BHF plotting theme}
\usage{
bhf_theme(textsize = 12, grid = FALSE, line = TRUE, map = FALSE)
}
\arguments{
\item{textsize}{Size of text (defaults to 12)}

\item{grid}{Boolean for whether grid lines should be drawn}

\item{line}{Boolean for whether line elements should be drawn}

\item{map}{Boolean for a map object}
}
\value{
ggplot2 style object
}
\description{
bhf_theme - The main BHF plotting theme
}
\examples{
theme <- bhf_style(line=TRUE, grid=FALSE reverse = TRUE)
}
