% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhf_theme.R
\name{bhf_theme_map}
\alias{bhf_theme_map}
\title{bhf_theme_map - The BHF plotting theme for maps}
\usage{
bhf_theme_map(textsize = 12)
}
\arguments{
\item{textsize}{Size of text (defaults to 12)}

\item{line}{Boolean for wether line elements should be drawn}

\item{grid}{Boolean for wether grid lines should be drawn}
}
\value{
ggplot2 style object
}
\description{
bhf_theme_map - The BHF plotting theme for maps
}
\examples{
theme <- bhf_style(line=TRUE, grid=FALSE reverse = TRUE)
}
