% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhf_theme.R
\name{create_palette}
\alias{create_palette}
\title{Creates a palette based on the BHF colours}
\usage{
create_palette(num_colours, num_steps = NULL, type = "discrete")
}
\arguments{
\item{num_colours}{Number of desired colours}

\item{num_cats}{Number of steps between colours}
}
\value{
Number of colours to use (integer)
}
\description{
Returns BHF palette based on the number of colours and categories.
}
\examples{
create_palette(5, 8)
}
