% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhf_theme.R
\name{get_colour_code}
\alias{get_colour_code}
\title{Get BHF Colour by Name}
\usage{
get_colour_code(x)
}
\arguments{
\item{x}{Name of the BHF colour (case insensitive)}
}
\value{
Hex code of the colour
}
\description{
Returns the hex code for a given BHF colour name.
}
\examples{
bhf_get_colour("blue")
}
