% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhf_theme.R
\name{scale_color_bhf}
\alias{scale_color_bhf}
\title{BHF Colour Scale}
\usage{
scale_color_bhf(palette = "bhf colours", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Colour palette - defaults to BHF standard colours}

\item{discrete}{Ensures the scale is discrete/continuous}

\item{reverse}{Reverses the palette order}
}
\value{
Palette object
}
\description{
Create a BHF colour scale
}
\examples{
temp1 <- scale_color_bhf("reds", reverse = TRUE)
}
