% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhf_theme.R
\name{scale_colour_bhf}
\alias{scale_colour_bhf}
\title{BHF Colour Scale}
\usage{
scale_colour_bhf(
  palette = "bhf colours",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Colour palette - defaults to BHF standard colours}

\item{discrete}{Ensures the scale is discrete/continuous}

\item{reverse}{Reverses the palette order}
}
\value{
Palette object
}
\description{
Create a BHF colour scale
}
\examples{
temp1 <- scale_colour_bhf("bhf colours", reverse = TRUE)
}
