% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhf_theme.R
\name{bhf_theme_map}
\alias{bhf_theme_map}
\title{bhf_theme_map - wrapper function for bhf_theme to support legacy bhf_theme_map function
Equivalent to bhf_theme_map(line=TRUE, grid=TRUE)}
\usage{
bhf_theme_map(textsize = 12)
}
\arguments{
\item{textsize}{Size of text (defaults to 12)}

\item{line}{Boolean for whether line elements should be drawn}

\item{grid}{Boolean for whether grid lines should be drawn}

\item{map}{Boolean for a map object}
}
\value{
ggplot2 style object
}
\description{
bhf_theme_map - wrapper function for bhf_theme to support legacy bhf_theme_map function
Equivalent to bhf_theme_map(line=TRUE, grid=TRUE)
}
\examples{
theme <- bhf_theme_map(grid=TRUE)
}
